requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-extensions==3.0.5
django-filter==2.3.0
django-ipware==3.0.1
django-model-utils==4.0.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
flake8==3.8.3
hgtk==0.1.3
idna==2.10
mccabe==0.6.1
mysqlclient==2.0.1
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
